Refactor resources, extract common fetch and create logic into trait
This should make it a lot easier to maintain and reduce duplicate code.
