+++ packages/Redmine/src/Version.php
@@ -2,18 +2,11 @@
 
 namespace Aedart\Redmine;
 
-use Aedart\Contracts\Redmine\Connection;
 use Aedart\Contracts\Redmine\Deletable;
-use Aedart\Contracts\Redmine\PaginatedResults as PaginatedResultsInterface;
-use Aedart\Contracts\Redmine\Resource;
 use Aedart\Contracts\Redmine\Updatable;
-use Aedart\Redmine\Exceptions\RedmineException;
-use Aedart\Redmine\Pagination\PaginatedResults;
 use Aedart\Redmine\Partials\Reference;
 use Aedart\Redmine\Relations\BelongsTo;
 use Carbon\Carbon;
-use JsonException;
-use Throwable;
 
 /**
  * Version Resource
@@ -41,6 +34,8 @@ class Version extends RedmineResource implements
     Updatable,
     Deletable
 {
+    use Concerns\ProjectDependentResource;
+
     /**
      * Open - version status (default)
      */
@@ -120,98 +115,6 @@ class Version extends RedmineResource implements
     public function resourceName(): string
     {
         return 'versions';
-    }
-
-    /**
-     * Fetch versions for the given project
-     *
-     * @param int|string|Project $project Project identifier or project instance
-     * @param callable|null $filters [optional] Callback that applies filters on the given Request {@see Builder}.
-     *                               The callback MUST return a valid {@see Builder}
-     * @param int $limit [optional]
-     * @param int $offset [optional]
-     * @param string|Connection|null $connection [optional] Redmine connection profile
-     *
-     * @return PaginatedResultsInterface<static>
-     *
-     * @throws Throwable
-     * @throws RedmineException
-     * @throws JsonException
-     */
-    public static function fetchForProject(
-        $project,
-        ?callable $filters = null,
-        int $limit = 10,
-        int $offset = 0,
-        $connection = null
-    ): PaginatedResultsInterface {
-        $endpoint = static::projectVersionEndpoint($project);
-
-        $resource = static::make([], $connection);
-        $response = $resource
-            ->applyFiltersCallback($filters)
-            ->limit($limit)
-            ->offset($offset)
-            ->get($endpoint);
-
-        return PaginatedResults::fromResponse($response, $resource);
-    }
-
-    /**
-     * Create a new version for given project
-     *
-     * @param int|string|Project $project Project identifier or project instance
-     * @param array $data
-     * @param string|Connection|null $connection [optional] Redmine connection profile
-     *
-     * @return static
-     *
-     * @throws JsonException
-     * @throws Throwable
-     */
-    public static function createForProject($project, array $data, $connection = null)
-    {
-        $endpoint = static::projectVersionEndpoint($project);
-        $resource = static::make([], $connection);
-
-        // Prepare payload
-        $prepared = $resource->prepareBeforeCreate($data);
-        $payload = [
-            $resource->resourceNameSingular() => $prepared
-        ];
-
-        // Perform create request
-        $response = $resource
-            ->request()
-            ->post($endpoint, $payload);
-
-        // Extract and (re)populate resource
-        return $resource->fill(
-            $resource->decodeSingle($response)
-        );
-    }
-
-    /**
-     * Returns a version url for given project
-     *
-     * @param int|string|Project $project Project identifier or project instance
-     *
-     * @return string
-     *
-     * @throws RedmineException|Throwable
-     */
-    public static function projectVersionEndpoint($project): string
-    {
-        $id = $project;
-        if ($project instanceof Resource) {
-            $id = $project->id();
-        }
-
-        if (!isset($id)) {
-            throw new RedmineException('Unable to resolve project version endpoint resource url');
-        }
-
-        return Project::make()->endpoint($id, 'versions');
     }
 
     /*****************************************************************
